+++ libstrawberry/src/crypto/symmetric/salsa20.h
@@ -64,8 +64,8 @@ typedef union sb_crypto_salsa20_ctx {
 extern "C" {
 #endif
 
-	SBAPI void sb_crypto_salsa20_init_ex(sb_crypto_salsa20_ctx_t *ctx, uint8_t bits, void *key, void *nonce, void *constant);
-	SBAPI void sb_crypto_salsa20_init(sb_crypto_salsa20_ctx_t *ctx, uint8_t bits, void *key, void *nonce);
+	SBAPI void sb_crypto_salsa20_init_ex(sb_crypto_salsa20_ctx_t *ctx, uint16_t bits, void *key, void *nonce, void *constant);
+	SBAPI void sb_crypto_salsa20_init(sb_crypto_salsa20_ctx_t *ctx, uint16_t bits, void *key, void *nonce);
 	SBAPI sb_bool_t sb_crypto_salsa20_clear(sb_crypto_salsa20_ctx_t *ctx);
 	SBAPI void sb_crypto_salsa20_nonce_increment(sb_crypto_salsa20_ctx_t *ctx);
 	SBAPI void sb_crypto_salsa20_nonce_decrement(sb_crypto_salsa20_ctx_t *ctx);
